requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
cryptography==41.0.7
assemblyai==0.17.0
openai==1.3.0
pytest==7.4.3
pytest-flask==1.3.0
Werkzeug==3.0.1